requirements.txt
dj-database-url 
gunicorn
django-widget-tweaks
djangorestframework
psycopg2-binary
whitenoise
drf-yasg